Implement YouTube design in Blender
Summary:
This implements the latest YouTube design specifications.
Some parts were not specified, so I made those up
(like the size of the leaf logo),
but the whole thing is parameterized
so we can change it easily.

Test Plan:
Open `youtube.blend`, and, in the console, execute

    batch.render_batch(youtube.YouTubeRenderer, "youtube.json")

Since you probably don't want to go to all that trouble,
here's the output instead:

{F167351}

{F167352}

{F167353}

Reviewers: desmond, emily, michelle

Reviewed By: michelle

Differential Revision: https://phabricator.khanacademy.org/D20313