Implement general-purpose thumbnails in Blender
Summary:
These are the normal thumbnails for videos, articles, and exercises.
The node setup for this is really simple:
{F167373}
The Python setup is also simple;
it's the same as the YouTube thumbnails, but there's no text.

I also added another test image that's probably more representative.

Again, we can in-person review this.

Test Plan:
Open `general.blend` and, in the console, execute

    batch.render_batch(general.GeneralRenderer, "general.json")

Then check that the output matches the following:

{F167364}

{F167365}

{F167366}

{F167367}

Reviewers: emily, michelle

Reviewed By: michelle

Differential Revision: https://phabricator.khanacademy.org/D20316